requirements.txt
absl-py==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
bidict==0.23.1
blinker==1.8.2
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
dm-tree==0.1.8
dnspython==2.6.1
etils==1.3.0
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
flatbuffers==24.3.25
flax==0.7.2
gast==0.4.0
gevent==24.2.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint==0.2.3
packaging==23.2
pandas==2.0.3
pillow==10.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.34
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-decision-forests==1.5.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflowjs==4.20.0
tensorstore==0.1.45
termcolor==2.4.0
tf-keras==2.15.0
threadpoolctl==3.5.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.1.1
zipp==3.18.2
zope.event==5.0
zope.interface==7.0.3
